Update index.html
Removed the packages due to the Heruku Postgres Database charging rates - hobby free has been dropped to make way for a base price  database which my client decided to drop the database idea.  She decided it was not worth the extra money at this stage.  So I have adapted it to exclude the packages but they can be easily reincorporated by reinstating and paying for the basic postgres.